backend/core/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.0.5
attrs==24.2.0
backcall==0.2.0
bleach==6.1.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==43.0.1
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dj-rest-auth==6.0.0
Django==5.1
django-admin-soft-dashboard==1.0.11
django-allauth==0.61.1
django-cors-headers==4.4.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-extensions==3.1.5
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==0.8.3
ffmpy==0.3.0
frozenlist==1.4.1
google-api-core==2.20.0
google-api-python-client==2.147.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
gotrue==2.8.1
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
hyperframe==6.0.1
idna==3.3
image==1.5.33
imageio==2.35.1
imageio-ffmpeg==0.4.5
ipython==7.0.1
jedi==0.18.1
kaggle==1.6.17
Markdown==3.7
matplotlib-inline==0.1.3
moviepy==1.0.3
multidict==6.1.0
numpy==2.1.0
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==24.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.4.0
postgrest==0.16.11
proglog==0.1.9
prompt-toolkit==2.0.10
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
Pygments==2.11.2
PyJWT==2.9.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2021.3
realtime==2.0.2
requests==2.27.1
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.24.0
simplegeneric==0.8.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.2
stack-data==0.2.0
storage3==0.7.7
StrEnum==0.4.15
supabase==2.7.4
supafunc==0.5.1
text-unidecode==1.3
tqdm==4.63.1
traitlets==5.1.1
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.9
uuid==1.30
wcwidth==0.2.5
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
whitenoise==5.3.0
wsproto==1.2.0
yarl==1.11.1
